add moving cloud ui
